fix(mining): 素通しプロパティを畳みテストを規約行数へ分割する
- EarnItemGuids を { get; private set; } へ畳む（MapObject/Outcrop）
- MiningFocusStateTest を土台・従来文言・名前つきの3ファイルへ分割
- 採掘テスト共有のリフレクション補助を MiningTestReflection へ集約
- 取得物なしPickUpの従来文言と、名前/ツールの並び順取り違え検出を追加
