+++ moorestech_client/Assets/Scripts/Client.Game/InGame/Map/MapObject/MapObjectGameObject.cs
@@ -28,8 +28,7 @@ namespace Client.Game.InGame.Map.MapObject
         // Targets that need no tool return an empty recommendation, so share one instance instead of allocating
         private static readonly List<ItemId> EmptyToolItemIds = new();
 
-        private IReadOnlyList<Guid> _earnItemGuids = Array.Empty<Guid>();
-        public IReadOnlyList<Guid> EarnItemGuids => _earnItemGuids;
+        public IReadOnlyList<Guid> EarnItemGuids { get; private set; } = Array.Empty<Guid>();
 
         public bool IsDestroyed { get; private set; }
         public int CurrentHp { get; private set; }
@@ -83,7 +82,7 @@ namespace Client.Game.InGame.Map.MapObject
 
             // 取得物はマスタ確定時に1度だけ拾う
             // Resolve the yields once the master is settled
-            _earnItemGuids = MapObjectMiningPresentation.GetEarnItemGuids(MapObjectMasterElement);
+            EarnItemGuids = MapObjectMiningPresentation.GetEarnItemGuids(MapObjectMasterElement);
 
             
             if (mapObjectInfo.IsDestroyed)
