+++ moorestech_client/Assets/Scripts/Client.Tests/Mining/MiningFocusStateEarnItemNameTest.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4abab320a14b54c36a171b1c8b0fdbfd
