+++ moorestech_client/Assets/Scripts/Client.Tests/Mining/MiningFocusStateTestFixture.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 426e6bab012364c819d02ec72f0c4582
