+++ moorestech_client/Assets/Scripts/Client.Tests/Mining/MiningTestReflection.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 341113e61aef8469fa23aa1f935e9c62
